requirements.txt
autoattack==0.1
einops==0.8.0
ftfy==6.2.0
h5py==3.9.0
numpy==1.24.3
opencv_python==4.9.0.80
Pillow==9.4.0
Pillow==10.3.0
pycountry==23.12.11
regex==2022.7.9
scipy==1.13.0
setuptools==68.0.0
timm==0.9.16
torch==2.1.2+cu118
torchvision==0.16.2+cu118
tqdm==4.65.0
ttach==0.0.3
